refactor: extract core extension service hooks
